Refactor: TimelineBuilderのDefinitionを分離し、処理を共通化
