+++ Assets/Project/Scenes/Battle/Scripts/Model/Definitions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5c1a39a171118314fa50efb7296894f5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
