+++ Assets/Project/Scenes/Battle/Scripts/Model/Definitions/AnimationTrackDefinition.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: fa6e32d073b13954ebaaffed593eb62a
