+++ Assets/Project/Scenes/Battle/Scripts/Model/Definitions/Project.Scenes.Battle.Model.Definitions.asmref
@@ -1,0 +1,3 @@
+{
+    "reference": "GUID:6bf788305a2b66346961b65504a3e059"
+}
