+++ Assets/Project/Scenes/Battle/Scripts/Model/Definitions/Project.Scenes.Battle.Model.Definitions.asmref.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b2ffd94900fa5ca4886b65be91829382
+AssemblyDefinitionReferenceImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
