+++ Assets/Project/Scenes/Battle/Scripts/Model/Definitions/SignalTrackDefinition.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 70c94fd4366fb4d4ba42b7a732882d53
