+++ Assets/Project/Scenes/Battle/Scripts/Model/Definitions/TrackDefinitionBase.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 5d453cd6d9745724bb0f1acb1ee61af9
